AzContainerTest1-main/event-job/requirements.txt
azure-storage-queue==12.8.0
